+++ packages/fresh/lib/fresh.dart
@@ -1,1 +1,2 @@
 export 'src/fresh.dart';
+export 'src/token.dart';